Fix one crash bug. Also a small code reorg.
With the new JSON and Java Event work in logstash core, we no longer use
the built-in JSON library, so it no longer makes sense to catch
exceptions for a class that may not be loaded (JSON::GeneratorError).
Other plugins (rabbitmq, for example), rescues StandardError here, so I
have copied that.

The `on_event` callback was also moved to a method `send_to_redis`.