Support for REDIS priority queues
This commit adds support for ZSET queues in REDIS with data_type 'list'.
Batch is supported.

Priority is based on a field of the message containing a number.
If the field is absent or not a number a default priority is then used.

It adds the following options to the configuration:
* **priority**: *boolean*, default false, enable priority mode
* **priority_field**: *string*, default '@timestamp', Priority field to
use, if field doesn't exist, priority will be priority_default
* **priority_default**: *number*, default '0', Default priority when
priority field is not found in the event

Usage example:
    output {
    	redis {
		host => ["127.0.0.1"]
		port => 6379
		data_type => "list"
		key => "syslog"
		priority => true
		priority_field => "epoch"
	}
    }