add support for the LPUSH command as a list operation
The LPUSH command inserts at the head of the list, as opposed to RPUSH
which inserts at the head of the list.

The default list operation remains RPUSH. However, LPUSH may be set via
the list_operation configuration option.

While here, fix documentation about the minimum version for RPUSH/LPUSH
support, which is 1.0.0 according to Redis documentation.